+++ app/templates/pages/organisation.html
@@ -7,10 +7,14 @@
 {% block content %}
   <a href="/" class="govuk-back-link">Back</a>
 
-  <h1 class="govuk-heading-xl">Owners</h1>
+  <h1 class="govuk-heading-xl">Find a GitHub Repository Owner</h1>
 
   <div class="govuk-grid-row">
-    <div class="govuk-grid-column-full">
+    <div class="govuk-grid-column-one-half">
+      <h3 class="govuk-heading-l">Repositories by Owners</h3>
+      {{ pie_chart | safe }}
+    </div>
+    <div class="govuk-grid-column-one-half">
       <h3 class="govuk-heading-l">Repositories by Relationship Type</h3>
       {{ pie_chart_of_admin_access | safe }}
     </div>
@@ -34,21 +38,33 @@
             Apply filters
           </button>
           <div class="govuk-form-group">
-            <label class="govuk-label" for="sort">
-              Owner
-            </label>
-            <select class="govuk-select" id="owner" name="owner">
-              {% for owner in owners %}
-                <option value="{{ owner }}" {% if owner == selected_owners %} selected {% endif %}>{{ owner }}</option>
-              {% endfor %}
-              <option value="NO_OWNER"  {% if "NO_OWNER" == selected_owners %} selected {% endif %} >No Owner</option>
-            </select>
-          </div>
-          <div class="govuk-form-group">
             <label class="govuk-label govuk-label--m" for="repository-name">
               Repository Name
             </label>
             <input class="govuk-input" id="repository-name" name="repository-name" type="text" value="{{ repository_name }}">
+          </div>
+          <div class="govuk-form-group">
+            <fieldset class="govuk-fieldset">
+              <legend class="govuk-fieldset__legend govuk-fieldset__legend--m">
+                Owner
+              </legend>
+              {% for owner in owners %}
+                <div class="govuk-checkboxes govuk-checkboxes--small" data-module="govuk-checkboxes">
+                  <div class="govuk-checkboxes__item">
+                    <input class="govuk-checkboxes__input" id="{{ owner }}" name="owner" type="checkbox" value="{{ owner }}" {% if owner in selected_owners %} checked {% endif %}>
+                    <label class="govuk-label govuk-checkboxes__label" for="{{ owner }}">
+                      {{ owner }}
+                    </label>
+                  </div>
+              {% endfor %}
+              <div class="govuk-checkboxes govuk-checkboxes--small" data-module="govuk-checkboxes">
+                <div class="govuk-checkboxes__item">
+                  <input class="govuk-checkboxes__input" id="NO_OWNER" name="owner" type="checkbox" value="NO_OWNER" {% if "NO_OWNER" in selected_owners %} checked {% endif %}>
+                  <label class="govuk-label govuk-checkboxes__label" for="NO_OWNER">
+                    No Owner
+                  </label>
+                </div>
+            </fieldset>
           </div>
           <div class="govuk-form-group">
             <fieldset class="govuk-fieldset">
@@ -67,20 +83,6 @@
               {% endfor %}
             </fieldset>
           </div>
-          <div class="govuk-form-group">
-            <fieldset class="govuk-fieldset">
-              <legend class="govuk-fieldset__legend govuk-fieldset__legend--m">
-                Options
-              </legend>
-              <div class="govuk-checkboxes govuk-checkboxes--small" data-module="govuk-checkboxes">
-                <div class="govuk-checkboxes__item">
-                  <input class="govuk-checkboxes__input" id="exclude_other_relationships_where_admins_exist" name="exclude_other_relationships_where_admins_exist" type="checkbox" value="true" {% if exclude_other_relationships_where_admins_exist %} checked {% endif %}>
-                  <label class="govuk-label govuk-checkboxes__label" for="exclude_other_relationships_where_admins_exist">Exclude Other Relationships When Admins Exist</label>
-                </div>
-              </div>
-            </fieldset>
-          </div>
-
         </div>
       </div>
     </div>
