+++ app/templates/pages/owner_dashboard.html
@@ -1,0 +1,40 @@
+{% extends "components/base.html" %}
+
+{% block pageTitle %}
+    Home
+{% endblock %}
+
+{% block content %}
+  <a href="/" class="govuk-back-link">Back</a>
+
+  <h1 class="govuk-heading-xl">{{ owner }} Dashboard</h1>
+  <p class="govuk-heading-m">Total Repositories: {{ filtered_repositories_by_owner | length }}</p>
+  <p class="govuk-heading-m">Admin Access: {{ repositories_with_admin_access | length }}</p>
+
+
+  <table class="govuk-table" data-module="moj-sortable-table">
+    <thead class="govuk-table__head">
+      <tr class="govuk-table__row">
+        <th scope="col" class="govuk-table__header" aria-sort="ascending">Name</th>
+        <th scope="col" class="govuk-table__header" aria-sort="none">Relationship</th>
+      </tr>
+    </thead>
+
+    <tbody class="govuk-table__body">
+      {% for repository in filtered_repositories_by_owner %}
+      <tr class="govuk-table__row">
+        <td class="govuk-table__cell"><a href="https://github.com/ministryofjustice/{{ repository["name"] }}" rel="noreferrer noopener" target="_blank" class="govuk-link">{{ repository["name"] }}</a></td>
+        <td class="govuk-table__cell">
+          {% for repository_owner in repository["owners"] %} 
+            {% if repository_owner["owner_name"] == owner %} 
+              {{ repository_owner["relationship_type"] }}
+            {% endif %} 
+          {% endfor %}
+        </td>
+      </tr>
+      {% endfor %}
+    </tbody>
+  </table>
+
+
+{% endblock %}
